requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0

# OCR related
pytesseract>=0.3.8
pdf2image>=1.16.0
easyocr>=1.5.0
Pillow>=9.0.0
opencv-python>=4.5.5

# NLP related
spacy>=3.4.0
transformers>=4.18.0
nltk>=3.6.7
flair>=0.11.0

# Pipeline and utilities
tqdm>=4.62.0
joblib>=1.1.0
pydantic>=1.9.0

# Notebook dependencies
jupyter>=1.0.0
ipywidgets>=7.7.0
